+++ src/platform_paths.h
@@ -3,6 +3,10 @@
 
 #include <stddef.h>
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 void platform_paths_init(const char *argv0);
 
 void platform_path_join_resource(char *dest, const char *relative, size_t size);
@@ -11,6 +15,7 @@ void platform_path_join_support(char *dest, const char *relative, size_t size);
 void platform_path_global_config(char *dest, size_t size);
 void platform_path_machine_config(char *dest, size_t size, const char *config_name);
 void platform_path_configs_dir(char *dest, size_t size);
+void platform_path_drives_dir(char *dest, size_t size);
 void platform_path_hostfs_dir(char *dest, size_t size);
 void platform_path_podules_user_dir(char *dest, size_t size);
 void platform_path_podules_bundle_dir(char *dest, size_t size);
@@ -19,5 +24,9 @@ int platform_path_find_rom_path(char *dest, const char *relative, size_t size);
 
 void platform_paths_init_test(const char *support, const char *resources);
 void platform_paths_reset(void);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif